Dungeon Procedural Gen
Created a separate scene to work on procedural generation. Created test prefabs and scripts to generate the dungeon. Currently the dungeon is generating rooms, however the connections aren't always correct. There is also an issue with how many rooms are being generated, sometimes it generates more rooms than was is to be expected(maxrooms) and sometimes less, which is expected to occur occasionally.
Also reloaded VSCode into the project, as well as updating the visual studio editor to allow vscode to work.