Working on dungeon Proc
In the middle of restructuring the data structure used by the rooms. Right now the rooms wont be connected properly. I need to copy the room connections after instantiating a room. Currently the room that gets instantiated wont have a reference to the rooms we want them connected to, and the referenced rooms wont have a reference to the room. Needs a function to copy the connections from the check room to the copy room