Made Room #1 and Hall#1 with all possible outcomes. Created Origin Room#1
Created a couple rooms and all possible room outcomes as well. Still need to add functionality to the chests, as well as make them randomly appear throughout dungeon.
Also fixed functionality of walking under the bridge in room 1. Previously enemies could appear on the bridge if they were underneath when the player crossed, as well as appear below if the player went underneath.