Added Rooms, Fixed Triggers, Load Handler
Added a room that will lead to the boss fight. The door begins locked and require a key. If the player enters he is moved far off the map, which is where the boss room will be located.
Added a load handler. The load handler will keep the player tracked from scene to scene. This allows the game to keep track of damage and other attribute changes made to the player. To achieve this I had to alter the health, camera and player script.